backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
python-dotenv==1.0.1
openai==1.58.1
anthropic==0.42.0
requests==2.32.3
httpx==0.28.1
google-generativeai==0.8.3
websockets==14.1
dashscope==1.20.11
ddgs==9.10.0
